requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
PyMySQL==1.1.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
python-dotenv==1.0.1
pytest==8.1.1